order/user_orders/utils.py
import jwt

# Ключ и алгоритм для дешифровки токена
SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
ALGORITHM = "HS256"


def get_user_id(token: str):
    """Получаем id пользователя из токена."""
    payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    user_id = payload.get("sub")
    return user_id
